requirements.txt
tornado==4.4.3
Pillow==4.1.0
